Create EditPostPage for Edit Post test
